lambda/requirements.txt
alpaca-py
requests
boto3

